R39-DECOMP-VIEWER ⑤: the project-browser panel out — app.ts 3,953 -> 3,751
buildPanels + buildSelSets, 216 contiguous lines, one concern. buildSelSets has
exactly one call site and it is inside buildPanels.

SCOPE IS SMALLER THAN APPROVED, DELIBERATELY. The plan was ~318 lines including
railGroup, distributeToolGroups, clearDistributed and refreshFederation. Every
call site of the first three is inside buildToolsPanel — they are its helpers,
so extracting them without their caller means exporting something app.ts
immediately re-imports to hand back to the function that stayed. That is churn
with a line-count reduction attached, which the ratchet would record as
progress. refreshFederation is worse: 8 call sites across the file, one already
threaded into qaSection. Deferred with reasons rather than taken for the number.

THREE read-write captures in 216 lines, which settles the reword. discTree and
colorMode were known; tsc found a third, spatialElements, which this code
REBUILDS (`spatialElements = elements.map(...)`) and A29-SPATIAL-SELECT reads on
every pick. All three cross as refs with ownership unmoved. The discipline ref
is a get/set view over the original `let`s, so NOT ONE READ SITE IN app.ts
CHANGES — disciplineOfClass and the colour lookup still read the variables.

Cost stated rather than hidden: 13 of 216 lines differ, against 2 in the
analyse slice. 11 are the ref rewrites; 2 are buildSelSets gaining an explicit
deps parameter, because it is a SIBLING function and cannot see the destructure
inside buildProjectPanels — passing deps beat hoisting them to module scope,
which would have been a hidden global. Verified by diffing the moved body
against `git show HEAD:app.ts` at 1388..1603: 216 lines in, 216 out, and every
differing line accounted for above.

MY OWN CENSUS UNDER-COUNTED AND TSC CAUGHT IT. seam.py matched only
declaration-shaped patterns, so it missed `$`, buildTree, spatialElements,
fitToItems, loadSelSets and askText. The tool's population was narrower than
its claim — the same defect this repo keeps finding in gates, this time in my
instrument for finding it. tsc enumerated all six; a census is a starting list,
not an authority.

It also caught me over-removing an import: I deleted the whole
selectionSetsStore line when only two of its four names were unused.

WHAT TSC DOES NOT GATE, AND THIS IS THE HONEST LIMIT: a by-value capture
compiles cleanly. tsc gates the THREADING (a missed dep is a compile error) but
cannot see ref-vs-value, which is why that rule is a discipline written into the
directions rather than a check. Dropping a dep from the interface fails with 2
errors; converting a ref to a value fails with none.

toolsSplit.test.ts needs no change: it reads TOOL SECTION sources and this block
contains none.

1307/1307 web green (116 files), tsc + eslint clean, test_file_sizes OK.
